resources/deployment/requirements.txt
# Lecture 83 - Deployment Requirements
# Python 3.9+

# Deep Learning
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0

# API Development
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6

# RAG and NLP
sentence-transformers==2.2.2
transformers==4.35.0
langchain==0.0.340
faiss-cpu==1.7.4
chromadb==0.4.18

# UI Frameworks
gradio==4.7.1
streamlit==1.28.1
streamlit-drawable-canvas==0.9.3

# Data Science
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.2

# Utilities
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0

# Development
jupyter==1.0.0
notebook==7.0.6
ipywidgets==8.1.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Docker
docker==6.1.3
